Add oauth token support (#5)
* Add ability to populate client from OAuth

TS has added (in Alpha) OAuth client support for generating scoped dynamic API
tokens. This adds support for specifying yor long lived OAuth client ID
and secret and subsequently using that to populate the TS client's
Authentication with the returned scoped token.

The Client's default behavior is still to take an API key directly in
the contructor, but if that value is None it will still work fine. You
can either populate the client directly from the OAuth token method by
setting 'client_embed' to True, or you can simply take the returned
object

* Switch client_update to be True by default

I figure if you're calling the OAuth clien code it's because you want to
use the token, so it makes more sense to update the client with that
info by default, but still allow the option of not doing so in case
someone needs that

* Update package version

Co-authored-by: Kevin Bringard <kevin.bringard@phreesia.com>