Add network flow log endpoint and re-organize examples (#6)
Tailscale added (as beta) network flow logs. If you enable this you can
now query those logs via the agent here. See the updated examples for
how to do this.

It's also worth noting: as of this writing there is no OAuth scope
specific to logging endpoints, so if you're using OAuth you need to
grant the "All" permission to the client. I have opened a case with
Tailscale to ask if/when they'd have a timeline on getting that added.

Please note the get_logs client method is currently aliased, but will be
deprecated in a future version. Please use get_audit_logs moving forward

Co-authored-by: Kevin Bringard <kevin.bringard@phreesia.com>